find code that doens't compile

+++ src/main/java/com/coderscampus/springwise/service/StudentService.java
@@ -3,6 +3,8 @@ package com.coderscampus.springwise.service;
 import java.util.List;
 import java.util.Optional;
 
+import javax.management.RuntimeErrorException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,27 +13,29 @@ import com.coderscampus.springwise.repository.StudentRepository;
 
 @Service
 public class StudentService {
-	
+
 	@Autowired
 	private StudentRepository studentRepo;
-	
-	// before: if (isValidNewStudent(student) || isValidStudentUpdate(student)) -> save student)
+
+	// before: if (isValidNewStudent(student) || isValidStudentUpdate(student)) ->
+	// save student)
 	public Student save(Student student) {
 		if (isValidNewStudent(student)) {
 			System.out.println("This is isValidNewStudent" + student);
-			return studentRepo.save(student); 
+			return studentRepo.save(student);
 		}
-		if(isValidStudentUpdate(student) ) {
+		if (isValidStudentUpdateOrDelete(student)) {
 			System.out.println("This is isValidStudentUpdate" + student);
 			return studentRepo.save(student);
 		}
 		return null;
 	}
 
-	boolean isValidStudentUpdate(Student student) {
+	boolean isValidStudentUpdateOrDelete(Student student) {
 		Optional<Student> existingStudent = studentRepo.findById(student.getId());
-
-		if (existingStudent.isPresent() && existingStudent.get().getUid() != null && existingStudent.get().getUid().equals(student.getUid())) {
+		Here is where we work
+		if (existingStudent.isPresent() && existingStudent.get().getUid() != null
+				&& existingStudent.get().getUid().equals(student.getUid())) {
 			System.out.println("This is existingStudent" + student);
 			return true;
 		}
@@ -45,7 +49,7 @@ public class StudentService {
 			System.out.println("These are the students " + students);
 			return false;
 		}
-		return student.getId() == 0; 
+		return student.getId() == 0;
 	}
 
 	public List<Student> findAll() {
@@ -58,8 +62,27 @@ public class StudentService {
 	}
 
 	public boolean delete(Student student) {
-		studentRepo.delete(student);
-		return false; // DN added this to allow testDeleteIfAllowed() to clear red
+
+		try {
+			if (isValidStudentUpdateOrDelete(student)) {
+
+				studentRepo.delete(student);
+
+				Optional<Student> user = studentRepo.findById(student.getId());
+				boolean foundUser = user.isPresent();
+				if (foundUser) {
+					throw new RuntimeErrorException(null, "User was not deleted");
+				}
+			} else {
+				return false;
+			}
+		} catch (Exception e) {
+			// TODO: handle exception
+			System.err.println(e);
+			return false;
+		}
+
+		return true;
 	}
 
 }
